% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{densifyMat}
\alias{densifyMat}
\title{Add vertices to a linear path to ensure curvature}
\usage{
densifyMat(x, minm = 1852 * 10)
}
\arguments{
\item{x}{matrix of coordinates, long/lat}

\item{minm}{default distance in metres on a great cicle}
}
\value{
matrix
}
\description{
This function takes a matrix of coordinates in longlat that define a pathy, and returns a denser matrix with more vertices based on a minimum allowed distance.
}
\details{
The function assumes each segment is a great circle, on the Vincenty ellipsoid \code{\link[geosphere]{distVincentyEllipsoid}}.
}
