% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/division.r
\name{division}
\alias{division}
\title{Define polygon object of CCAMLR divisions.}
\usage{
division(name = "58.4.1", intsct = NULL)
}
\arguments{
\item{name}{for future use, currently ignored}

\item{intsct}{optional poylgon object to use for cropping via intersection}
}
\value{
SpatialPolygonsDataFrame
}
\description{
Currently only builds objects for 58.4.1, ensuring they are topologically clean and with sufficient density for fishing impact analysis.
}
\examples{
\dontrun{
#cm <- readOGR(file.path(dp, "data", "add_ver6"), "cst01_polygon")
#cmu <- gUnionCascaded(cm)
library(aceecostats)
subarea <- division(intsct = aes_region)
}
}

